Update new Language Português-Brasileiro
There are no confirmation of a Brazilian version of Coriolis, even though we have many Free League/Mophidius games published here, so I took the liverty of making a version of the system basics.  Some translation decisions were based on how we already translate RPGs to portuguese like HIt Points (witch are "Puntos de golpe" in Spanish) were translated to "Pontos de Vida"(life points) instead of "Pontos de Acerto" (the literal translation to hit points) because it's the usual for 99% of the RPGs published here.
"Strength" and "Force" are also the same word (so is the Star Wars Force), so as the Year Zero concept of Strength atribute is like a sum of Strenght and Constitution (they even use Physique in Vaesen) I used "Vigor" to not get people confused about witch one is the Skill and Witch one is the Atribute.
As for Data Djinn portuguese has a problem that Djinn, Genie and Genius are the same word (Gênio) so I went for middle eastern original "Jinn" to not be confused as somone that's just ingenious with computers.
Though this might change when some Brazilian publisher decides to publish the game here, I hope this might be of some use for any gm hoping to run a game for their players who don't speak english or just want a more naturalistic playthrough.