in 0.8.x having the actor object in the rollData caused a serialization problem. Turns out I didn't need actor reference in any case.
More correctly, rolls should be actor agnostic. If I wish to link an actor in the future, the proper thing to do is to resolve th roll and then take the result and mix with actor data, OR use actor data to change modifiers. Instead of having rolls be 'aware' of actors.