fix: pushing rolls as GM or a player
previously, when the custom flag was set on a chat message it was also serializing the Roll object. I wasn't able to hunt down the specifics, but it seemed to be an issue with the doc updater trying to change a value on a roll via sync when the value is only a getter in v11.

Rather than send the Roll object entirely, I instead pull the roll object out and just pass it in as a parameter as needed in the roll functions. Then the results data has coriolis specific data that can serialize properly across the wire.