changes the feat Roll Visibility to an option
Many people already use different styles of ways for the roll-visbilites (like the module dice tray). The last implemented feat is a great idea but some people prefer to handle that with modules or the foundry-default way. The problem is that this feat overwrites all other modules.

This request makes it an option where the GM can choose if he wants to use it or not.